requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlmodel==0.0.14
pytest==7.4.3
pytest-asyncio==0.23.2
